+++ CHANGELOG.md
@@ -9,12 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### ✅ Added
 - `avatarViewStyle` under `ChatViewStyle` for customizing Navigation Right Bar Button Item avatar [#241](https://github.com/GetStream/stream-chat-swift/issues/241)
 - `logAssert(_:_:)` and `logAssertionFailure(_:)` functions added to `ClientLogger` [#231](https://github.com/GetStream/stream-chat-swift/issues/231).
-- `queryChannels` now returns unread count of each channel [#247](https://github.com/GetStream/stream-chat-swift/issues/247):
+- `queryChannels` now returns unread count of each channel, unrestricted by number of messages fetched [#247](https://github.com/GetStream/stream-chat-swift/issues/247):
   Example:
   ```swift
   Client.shared.queryChannels(filter: .currentUserInMembers, sort: [Sorting("has_unread", isAscending: false)]) { (result) in
       for response in result.value! {
-          print("Channel: \(response.channel.name!), unread count: \(response.messageReads.first(where: { $0.user == User.current })!.unreadMessageCount)")
+          print("Channel \(response.channel.name ?? "nil"), unread messages count: \(response.channel.unreadCount.messages)")
       }
   }
   ```